build/ci: Update Go to v1.24, golangci-lint to v1.64.5 and fix issues

+++ cmd/gen_sqlboiler_config/main.go
@@ -24,7 +24,7 @@ var sqlboilerConfig map[string]driverConfig
 type driverConfig struct {
 	DBName    string   `json:"dbname,omitempty"`
 	Host      string   `json:"host,omitempty"`
-	Port      uint16   `json:"port,omitempty"`
+	Port      uint32   `json:"port,omitempty"`
 	User      string   `json:"user,omitempty"`
 	Pass      string   `json:"pass,omitempty"`
 	Schema    string   `json:"schema,omitempty"`
